docs: add SynchronizedLineChart example in TypeScript (.tsx)
